runfix: wrapp image upload button with form and move it to separate file
