+++ src/script/page/message-list/InputBarControls/ControlButtons.tsx
@@ -18,12 +18,12 @@
  */
 
 import Icon from 'Components/Icon';
-import React, {useRef} from 'react';
+import React from 'react';
 import MessageTimerButton from '../MessageTimerButton';
 import {t} from 'Util/LocalizerUtil';
 import {Conversation} from 'src/script/entity/Conversation';
-import {Config} from '../../../Config';
 import {AssetUploadButton} from '../AssetUploadButton';
+import {ImageUploadButton} from '../ImageUploadButton/ImageUploadButton';
 
 export type ControlButtonsProps = {
   input: string;
@@ -54,10 +54,6 @@ const ControlButtons: React.FC<ControlButtonsProps> = ({
   onCancelEditing,
   onClickGif,
 }) => {
-  const acceptedImageTypes = Config.getConfig().ALLOWED_IMAGE_TYPES.join(',');
-
-  const imageRef = useRef<HTMLInputElement>(null!);
-
   const pingTooltip = t('tooltipConversationPing');
 
   if (isEditing) {
@@ -96,25 +92,7 @@ const ControlButtons: React.FC<ControlButtonsProps> = ({
             </li>
 
             <li>
-              <button
-                type="button"
-                aria-label={t('tooltipConversationAddImage')}
-                title={t('tooltipConversationAddImage')}
-                className="conversation-button controls-right-button no-radius file-button"
-                onClick={() => imageRef.current?.click()}
-                data-uie-name="do-share-image"
-              >
-                <Icon.Image />
-
-                <input
-                  ref={imageRef}
-                  accept={acceptedImageTypes}
-                  tabIndex={-1}
-                  id="conversation-input-bar-photo"
-                  onChange={({target: {files}}) => files && onSelectImages(Array.from(files))}
-                  type="file"
-                />
-              </button>
+              <ImageUploadButton onSelectImages={onSelectImages} />
             </li>
 
             <li>
